+++ tests/test_early_stopping.py
@@ -4,7 +4,6 @@
 
 import unittest
 from typing import List
-from unittest.mock import Mock
 
 import numpy
 import pytest
@@ -14,11 +13,10 @@ from torch.optim import Adam
 
 from pykeen.datasets import Nations
 from pykeen.evaluation import RankBasedEvaluator
-from pykeen.models import FixedModel, Model, TransE
+from pykeen.models import Model, TransE
 from pykeen.stoppers.early_stopping import EarlyStopper, EarlyStoppingLogic, is_improvement
 from pykeen.training import SLCWATrainingLoop
-from pykeen.typing import RANK_REALISTIC, SIDE_BOTH
-from tests.mocks import MockEvaluator
+from tests import cases
 
 
 class TestRandom(unittest.TestCase):
@@ -54,112 +52,6 @@ class TestRandom(unittest.TestCase):
                 )
 
 
-class LogCallWrapper:
-    """An object which wraps functions and checks whether they have been called."""
-
-    def __init__(self):
-        self.called = set()
-
-    def wrap(self, func):
-        """Wrap the function."""
-        id_func = id(func)
-
-        def wrapped(*args, **kwargs):
-            self.called.add(id_func)
-            return func(*args, **kwargs)
-
-        return wrapped
-
-    def was_called(self, func) -> bool:
-        """Report whether the previously wrapped function has been called."""
-        return id(func) in self.called
-
-
-class TestEarlyStopper(unittest.TestCase):
-    """Tests for early stopping."""
-
-    #: The window size used by the early stopper
-    patience: int = 2
-    #: The mock losses the mock evaluator will return
-    mock_losses: List[float] = [10.0, 9.0, 8.0, 9.0, 8.0, 8.0]
-    #: The (zeroed) index  - 1 at which stopping will occur
-    stop_constant: int = 4
-    #: The minimum improvement
-    delta: float = 0.0
-    #: The best results
-    best_results: List[float] = [10.0, 9.0, 8.0, 8.0, 8.0]
-
-    def setUp(self):
-        """Prepare for testing the early stopper."""
-        # Set automatic_memory_optimization to false for tests
-        self.mock_evaluator = MockEvaluator(
-            key=("hits_at_10", SIDE_BOTH, RANK_REALISTIC),
-            values=self.mock_losses,
-            automatic_memory_optimization=False,
-        )
-        nations = Nations()
-        self.model = FixedModel(triples_factory=nations.training)
-        self.stopper = EarlyStopper(
-            model=self.model,
-            evaluator=self.mock_evaluator,
-            training_triples_factory=nations.training,
-            evaluation_triples_factory=nations.validation,
-            patience=self.patience,
-            relative_delta=self.delta,
-            larger_is_better=False,
-        )
-
-    def test_initialization(self):
-        """Test warm-up phase."""
-        for epoch in range(self.patience):
-            should_stop = self.stopper.should_stop(epoch=epoch)
-            assert not should_stop
-
-    def test_result_processing(self):
-        """Test that the mock evaluation of the early stopper always gives the right loss."""
-        for epoch in range(len(self.mock_losses)):
-            # Step early stopper
-            should_stop = self.stopper.should_stop(epoch=epoch)
-
-            if not should_stop:
-                # check storing of results
-                assert self.stopper.results == self.mock_losses[: epoch + 1]
-                assert self.stopper.best_metric == self.best_results[epoch]
-
-    def test_should_stop(self):
-        """Test that the stopper knows when to stop."""
-        for epoch in range(self.stop_constant):
-            self.assertFalse(self.stopper.should_stop(epoch=epoch))
-        self.assertTrue(self.stopper.should_stop(epoch=self.stop_constant))
-
-    def test_result_logging(self):
-        """Test whether result logger is called properly."""
-        self.stopper.result_tracker = mock_tracker = Mock()
-        self.stopper.should_stop(epoch=0)
-        log_metrics = mock_tracker.log_metrics
-        self.assertIsInstance(log_metrics, Mock)
-        log_metrics.assert_called_once()
-        _, call_args = log_metrics.call_args_list[0]
-        self.assertIn("step", call_args)
-        self.assertEqual(0, call_args["step"])
-        self.assertIn("prefix", call_args)
-        self.assertEqual("validation", call_args["prefix"])
-
-    def test_serialization(self):
-        """Test for serialization."""
-        summary = self.stopper.get_summary_dict()
-        new_stopper = EarlyStopper(
-            # not needed for test
-            model=...,
-            evaluator=...,
-            training_triples_factory=...,
-            evaluation_triples_factory=...,
-        )
-        new_stopper._write_from_summary_dict(**summary)
-        for key in summary.keys():
-            assert getattr(self.stopper, key) == getattr(new_stopper, key)
-
-
 class TestEarlyStoppingLogic(unittest_templates.GenericTestCase[EarlyStoppingLogic]):
     """Tests for early stopping logic."""
 
@@ -188,7 +80,17 @@ class TestEarlyStoppingLogic(unittest_templates.GenericTestCase[EarlyStoppingLog
             self.assertEqual(stop, epoch >= 4)
 
 
-class TestEarlyStopperDelta(TestEarlyStopper):
+class TestEarlyStopper(cases.EarlyStopperTestCase):
+    """Tests for early stopping."""
+
+    patience: int = 2
+    mock_losses: List[float] = [10.0, 9.0, 8.0, 9.0, 8.0, 8.0]
+    stop_constant: int = 4
+    delta: float = 0.0
+    best_results: List[float] = [10.0, 9.0, 8.0, 8.0, 8.0]
+
+
+class TestEarlyStopperDelta(cases.EarlyStopperTestCase):
     """Test early stopping with a tiny delta."""
 
     mock_losses: List[float] = [10.0, 9.0, 8.0, 7.99, 7.98, 7.97]
